Fix reduced polycone
This fixes the ReducedPolycone utility for segment checks. Additional cleanup was done concerning passing vectors by value.
A test was added for the polycone RZ constructor, which was failing in CMS. The Vector destructor has been fixed for the composition case, where the same object may be deleted multiple times.

* [polycone] Passing vectors by reference rather than by value in ReducedPolycone

* [polycone] Add tolerance to ReducedPolycone line intersection checker.

* [polycone] Add RZ test case in the benchmarker.

* [polycone] Add critical error in the RZ constructor if the contour check fails.

* Fix Vector double delete of its data in case of composition.