[navigation] Global navigation tweaks and fixes.
Negative distance to daughters can be due to a push, and was previously ignoring the respective daughter - now the point is considered inside, preventing an extra relocation step and even errors in case of field propagation.

If the distance to exiting is shorter, navigation needs to find the deepest mother containing the point. Before, UnplacedContain was used, so if the point was on surface it was considered inside, then the relocator wrongly located to that parent, while the track was exiting.

Synchronized older navigators with these changes and removed some dead code for debugging.

* [navigation] Global navigation tweaks and fixes.

Negative distance to daughters can be due to a push, and was previously ignoring the respective daughter - now the point is considered inside, preventing an extra relocation step and even errors in case of field propagation.

If the distance to exiting is shorter, navigation needs to find the deepest mother containing the point. Before, UnplacedContain was used, so if the point was on surface it was considered inside, then the relocator wrongly located to that parent, while the track was exiting.

Synchronized older navigators with these changes and removed some dead code for debugging.