cleaning of the torus
This MR cleans some part of the torus. Some data structures were both in the UnplacedTorus2 class and in the torus struct. In the kernels, only the ones from the torus struct are used, therefore this MR is merely cleaning and not improving performance.
It also fixes some issue in the initialization of the bounding tube, where garbage values were observed.

* clean torus