RZ polycone fixes
Vector.h
- Vector operates on a C array, calling construction/destruction in place for the template elements. Several operations, like resizing, were calling in place destruction of elements when e.g. reducing the size w/o re-allocation of the array. In such case, the destructor of Vector and implicitly of its internal array was destructing automatically ALL elements, resulting in the double deletion of some, with unpredictible behavior.
- The current version replaces in place destruction with reset (destruction + default construction in place), so the destructor can never be called twice on the same slot content.
- TestVector was extended to thoroughly test features, including Vector<Vector<T>>

Cone/Polycone GPU construction fix
- In case of cones having rmin = rmax for a section, rmax is increased with `1e-7` to avoid numerical errors, and the original rmax is cached as a data member of the cone. The original value IS actually used in navigation of polycone sections to check if a cone or tube algorithm should be used. The `CopyToGPU` function was not rebuilding the polycone on device using the original values but the modified rmax, resulting in triggering the cone solver rather than the tube one, giving differences of the order `1e-7` compared to the host version.
- This correcly propagates the original cone Rmax values to GPU and fixes the cone GetRmax getters so that conversions to the surface model does not lead to the same issue.

* [test] Added a problematic CMS RZ polycone both in the polycone benchmarker and as unit test for polycones.

* [solids] Fix construction on GPU of cones/polycones with sections with rmin = rmax

In case of coned having rmin = rmax for a section, rmax is increased with 1e-7 to avoid numerical errors, and the original rmax is stored as data member of the cone. The original value IS actually used in navigation of polycone sections to check if a cone or tube algorithm should be used. The CopyToGPU function was not rebuilding the polycone on device using the original values, resulting in trigerring the cone solver rather than the tube one, giving differences of the order 1e-7 compared to the host version.

This correcly propagates the original cone Rmax values to GPU and fixes the cone GetRmax getters so that conversions to the surface model does not lead to the same issue.

* [benchmarker] Benchmarker prints the magnitude of the difference in distanced for verbosity > 2

* [solids] Small cleanup of ReducedPolycone

* [base] Cleaned possible double destruction of Vector elements.

Vector operates on a C array, calling construction/destruction in place for the template elements. Several operations, like resizing, were calling in place destruction of elements when e.g. reducing the size w/o re-allocation of the array. In such case, the destructor of Vector and implicitly of its internal array was destructing automatically ALL elements, resulting in the double deletion of some, with unpredictible behavior.

The current version replaces in place destruction with reset (destruction + default construction in place), so the destructor can never be called twice on the same slot content.

TestVector was extended to thoroughly test features, including Vector<Vector<T>>