Fix for bounding box overshooting in the BVH single-precision.
Observed failures in the previous algorithm of calculating the safe approach distance to the bbox. The current algorithm uses ULP (unit in the last place) for the distance to box value, using 10 ULP reduction of this distance to approach safely.

Added unit test for overshooting `TestOvershoot`

* Shorten approach distance to bbox with 10*ULP to be safer.

* Fix for bounding box overshooting in the BVH single-precision.

Observed failures in the previous algorithm of calculating the safe approach distance to the bbox. The current algorithm estimates max roundings using ULP (unit in the last place) for the distance to box value, considering 3 rounding operations of max 0.5 ULP each.