remove harmful truncation in NavigationTable
In constructing the Navigation table for NavTuple, the number of daughters was unnecessarily truncated to short.
This caused crashes in the full ATLAS geometry, as the ITKPixel detector has a volume with >160000 daughters. As the data is already stored in an array of `NavIndex_t = unsigned int`, the data storage is not affected.

A throw was added to avoid this issue in the future.

* remove harmful truncation