Move precision/navindex/uint into vegeom namespace
A couple of type aliases were in the global namespace. This moves Precision into Math.h, aliases it into the vgbrep namespace, and adds a `VECGEOM_PRECISION_NAMESPACE` feature macro so that downstream applications can compile with 'master' and future rc/dev releases as well as previous dev releases and v1.x.

* Fix CUDA build

* Add feature macro

* Add types to vgbrep namespace

* Move uint/navindex into namespace

* Move Precision into math.h in vecgeom namespace