Unify BVH and BVHSurf
This removes `BVHSurf` in order to avoid keeping two mirrored implementations of the same code. Both `BVH` and `BVHSurf` provide the same functionality although the implementations had diverged somewhat since they were introduced.

The function `AABB::IntersectInvDirApproach` was found to be sometimes returning points that fell inside the approached AABBs. In order to fix this, the error estimation was increased from 10 ULP to 500 ULP. This value seems to work consistently for both solids and surfaces.

Along with these changes `BVHNavigatorV` is also removed. This navigator was not being updated.

* Relax safety check

* Remove leftover comments

* Simplify check

* Move SetPointers

* Move Set function

* Revert to solids BVH ComputeSafety

* Remove BVHSurf

* Remove templation from ApproachNextDaughter, increase ComputeIntersectionInvDir error estimate from 10 to 500

* Move precision-templated ApproachNextDaughter from BVHSurf to BVH

* Move LevelLocate from BVHSurf to BVH

* Move ComputeSafety from BVHSurf to BVH

* Move harmless functions from BVHSurf to BVH

* Remove AABBSurf

* Remove BVHNavigatorV