CudaRdcUtils: Avoid no-undefined when constructing middle library
Also explicitly allow undefined symbols for rdc library (in case
--no-undefined becomes the default for linker)

The cuda_rdc middle library has by definition many unresolved symbols.
Those symbols are generated by nvlink whose result is only included in
the final library.

In order to neutralize a `-Wl,--no-undefined` set by the user

If the user request `-Wl,--no-undefined` to be applied/used when
generating a shared library, we must remove it for the middle
library generation (we could not find a flag that took priority/reverted
the `no-undefined` if/when it is passed to the linker).

This PR is independent of the VecGeom branch and once approved will be applied to master.

This is required for building within the LHCb context where we have:

```
set(CMAKE_SHARED_LINKER_FLAGS_INIT "-Wl,--as-needed -Wl,--no-undefined")
```

This patch means that for CMake project that create any RDC libraries,
and relies on setting `--no-undefined` via `CMAKE_SHARED_LINKER_FLAGS[_INIT]` then
`--no-undefined` will be disabled globally (as opposed to 'just' for the CUDA rdc
middle library).

* CudaRdcUtils: add missing cudart for non-cuda library that depends on a cuda library

* CudaRdcUtils: only use allow-shlib-undefined when available

* Use `-z undefs` when available

* CudaRdcUtils: use LINKER: for compiler independence

* CudaRdcUtils: remove dead code and debug code

* CudaRdc: Remove restoration of CMAKE_SHARED_LINKER_FLAGS

* CudaRdcUtils: Remove no-undefined from CMAKE_SHARED_LINKER_FLAGS.*

Also explicitly allow undefined symbols for rdc library (in case
--no-undefined becomes the default for linker)

The cuda_rdc middle library has by definition many unresolved symbols.
Those symbols are generated by nvlink whose result is only included in
the final library.

In order to neutralize a `-Wl,--no-undefined` set by the user

If the user request `-Wl,--no-undefined` to be applied/used when
generating a shared library, we must remove it for the middle
library generation (we could not find a flag that took priority/reverted
the `no-undefined` if/when it is passed to the linker).

* Upgrade CudaRdcUtils to v13