Fix macOS warnings
Fixes two new macOS warnings. A variable in a BVH constructor was uninitialized, and an if statement meant to support building the BVH for surfaces was actually unused.

The constructor now receives the list of AABBs and the number of children of a node from the caller. This still allows to use the constructor to build a BVH from surface AABBs if ever needed, and removes the unused if statement by making the caller retrieve the AABB list

* Fix macOS warnings