fix bug in LocatePointIn and add new LocatePointInNavState
There was an unnoticed bug in the `LocatePointIn` function calls both in the BVHNavigator and the LooperNavigator:

The function takes a point in the reference frame of the **parent** of the passed `vol`. Then, if `top` is true, it calls the `Inside()` of the `vol`, which internally transforms the point into the reference frame of `vol` and checks for the inside. However, the point itself inside the `LocatePointIn` is unchanged. Then, it is passed to the BVH. The BVH however, requires a point in the local reference frame of the volume `vol`. Thus, the transformation from the point in the reference frame of the parent of `vol` to the local reference frame of `vol` itself was missing, both in the BVHNavigator and LooperNavigator.

This most likely went unnoticed as this causes only an issue when the `vol` is not the world volume. 

The doxygen of the functions is added to clarify in which reference the passed point needs to live.

Furthermore, a new `LocatePointInNavState` is introduced: it takes a localpoint in the reference frame of the navState and locates it within the navState, finding the deepest navState that contains the point.

* improve doc

* fix usage of LocatePointIn in RelocatePoint after change of behaviour of LocatePointIn

* fix bug in LocatePointIn and add new LocatePointInNavState