Remove broken CUDA RNG to fix nvhpc compilation
The CUDA RNG in vecgeom causes build errors in nvhpc installations due to the different path used by curand. It also is unusable due to parallel thread semantics, and it's not used. Even the host RNG should be reexamined since it's not thread-safe.

Builds and all tests pass with cuda+gdml+surface.

* Empty

* Fix annotations

* Remove broken CUDA RNG