Fix RelocateToNextVolume
This MR is a followup to #1354, which missed another bug. `RelocateToNextVolume` was passing points in the coordinates of `vol` to `LocatePointIn`, instead of the parent of `vol`. To avoid further confusion, the cleaner and simpler function `LocatePointInNavState` is used. 

This leaves `LocatePointIn` unused in VecGeom, could be considered to be deprecated and deleted.

* fix usage of empty nav states

* reduce use of error prone LocatePointIn

* fixRelocateToNextVolume after bug fix in LocatePointIn