Fix issue affecting solids BVH
Follow up to https://gitlab.cern.ch/VecGeom/VecGeom/-/commit/21f12032f32898665622262515380169c9ff3860, the check for a positive distance before updating the safety in the BVH, needed for use with surfaces, caused unexpected navigation changes in the ATLAS detector geometry.
The check is disabled for now, ensuring that the solids model works correctly, and the code will be fixed for surfaces in the future.

In addition, the error overestimation used to estimate an approach to AABBs was found to be too large, causing false positives when checking for an intersection. This is fixed by using a large overestimation only after confirming an intersection. This is needed by the surface model in order to prevent too-large approaches which overshoot the surface.

* Fix issue affecting solids BVH